Splitting Users Files
Fixing the First Capital Letter

Cleaning the Image Service File

Fixing the Product Existence Check
