+++ package.json
@@ -32,7 +32,7 @@
     "build": "tsc && tsc-alias && cp -r src/docs dist/docs",
     "start": "node dist/index.js",
     "dev": "ts-node -r tsconfig-paths/register src/index.ts",
-    "dev:watch": "nodemon --watch src --ext ts,json --exec ts-node -r tsconfig-paths/register src/index.ts",
+    "dev:watch": "nodemon --watch src --watch swagger --ext ts,json,yaml,yml --exec ts-node -r tsconfig-paths/register src/index.ts",
     "test": "echo \"Error: no test specified\" && exit 1",
     "migrate": "ts-node -r tsconfig-paths/register src/infrastructure/database/run-migration.ts",
     "setup-db": "ts-node -r tsconfig-paths/register src/infrastructure/database/setup-database.ts"
