+++ src/application/interface/IFileStorageService.ts
@@ -1,10 +1,10 @@
 export interface IFileStorageService {
-    uploadFile(file: Express.Multer.File, userId: string, userType: 'customer' | 'restaurant_owner'): Promise<string>;
-    uploadProductFile(
+    UploadImageProfile(file: Express.Multer.File, userId: string, userType: 'customer' | 'restaurant_owner'): Promise<string>;
+    UploadProductImage(
         file: Express.Multer.File, 
         id: string, 
         type:  'restaurant_owner' | 'product',
         existingUrl?: string
     ): Promise<string>;
-    deleteFile(fileUrl: string): Promise<void>;
+    DeleteOldImage(fileUrl: string): Promise<void>;
 }
