+++ src/application/use-cases/product/CreateProductUseCase.ts
@@ -37,7 +37,7 @@ export class CreateProductUseCase {
         // Upload image if provided
         let imageUrl: string | null = null;
         if (image) {
-            imageUrl = await this.fileStorageService.uploadProductFile(image, uuidv4(), 'product');
+            imageUrl = await this.fileStorageService.UploadProductImage(image, uuidv4(), 'product');
         }
 
         const product: Product = {
@@ -54,13 +54,13 @@ export class CreateProductUseCase {
             updatedAt: new Date()
         };
 
-        await this.checkExistingByName(request);
+        await this.ExistingByNameAndRestaurantId(request, restaurantOwnerId);
         return await this.productRepository.create(product);
     }
-    private async checkExistingByName(request: CreateProductRequest): Promise<void> {
-        const productExists = await this.productRepository.checkExistingByName(request.productName);
+    private async ExistingByNameAndRestaurantId(request: CreateProductRequest, restaurantOwnerId: string): Promise<void> {
+        const productExists = await this.productRepository.ExistingByNameAndRestaurantId(request.productName, restaurantOwnerId);
         if (productExists) {
-            throw new Error('Product already exists with this name');
+            throw new Error('Product already exists with this name in your restaurant');
         }
     }
 }
