+++ src/domain/repositories/IProductRepository.ts
@@ -4,7 +4,7 @@ export interface IProductRepository {
     create(product: Product): Promise<Product>
     findById(id: string): Promise<Product | null>
     findByRestaurantOwnerId(restaurantOwnerId: string): Promise<Product[]>
-    checkExistingByName(productName: string): Promise<boolean>;
+    ExistingByNameAndRestaurantId(productName: string, restaurantOwnerId: string): Promise<boolean>;
     update(id: string, product: Partial<Product>): Promise<Product>
     delete(id: string): Promise<void>
 }
