+++ src/presentation/controller/Users/index.ts
@@ -1,8 +1,8 @@
-export * from './customerLogin'
-export * from './getRestaurantOwnerProfile'
-export * from './registerCustomer'
-export * from './registerRestaurantOwner'
-export * from './restaurantOwnerLogin'
-export * from './updateRestaurantLocation'
-export * from './uploadProfileImage'
-export * from './verifyEmail'
+export * from './Customer/CustomerLoginController'
+export * from './RestaurantOwner/GetRestaurantOwnerProfileController'
+export * from './Customer/RegisterCustomerController'
+export * from './RestaurantOwner/RegisterRestaurantOwnerController'
+export * from './RestaurantOwner/RestaurantOwnerLoginController'
+export * from './RestaurantOwner/UpdateRestaurantLocationController'
+export * from './UploadProfileImageController'
+export * from './VerifyEmailController'
